[citizen:scripting:lua] force pushed floating point values from Lua to be float, not default lua_Number (usually double)
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=e80f9d8bcf60851b10720a7c1f34332ffee74e05

+++ components/citizen-scripting-lua/src/LuaScriptRuntime.cpp
@@ -232,7 +232,7 @@ int Lua_InvokeNative(lua_State* L)
 			}
 			else
 			{
-				push(lua_tonumber(L, i));
+				push(static_cast<float>(lua_tonumber(L, i)));
 			}
 		}
 		// boolean
